# perf(serde_v8/de): implement SeqAccess size_hint